Don't show "hidden" fields in edit records table.
Moves all guards out of biosys-core into biosys-web shared.

Added Team Member guard that ensures users are team members when when editing records.

Added to Data Engineer guard to ensure DEs can only edit projects that are in the program that they are DE of.

Finally got logout menu on the right.