Added MediaManager component for viewing, uploading and deleting media.
MediaManager component added to edit-project and edit-dataset, which now have attached media.

Replacing growls with toast. This is now done for all management pages.

Using app-wide messageService, rather than component specific messages. This means when a component leaves (e.g. after saving) it can post a message to be displayed in a toast for the next screen. This is done for all management pages.