Added bounds to feature map to allow user to set bounds to zoom to.
Manage data also has map extent which will zoom the map to the extent of the dataset (if available) or the project bounds (if available).

Refactored the way projects and dataset are fetched in Manage Data.

Edit Records component will zoom to the record's location within 100 square km. If it's a new record it will use the extent of the dataset (if available) or the project bounds (if available).

Refactor of the way project / dataset and records fetched, along with record media and child records in Edit Records.

Updated leaflet version.